refactor: move linear imputation out of winddata
